Refactored Tests into their own Classes
Refactored tests from the Utilities.cs file into their own classes as the utilities class was getting long.
Also want to be able to quickly add additional tests and test in isolation.
