+++ src/Impl/Impl.csproj
@@ -47,8 +47,11 @@
     <Compile Include="BaseView.cs" />
     <Compile Include="Context.cs" />
     <Compile Include="FileSystemProvider.cs" />
+    <Compile Include="LogFind.cs" />
     <Compile Include="Logger.cs" />
+    <Compile Include="LogTime.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="LogLine.cs" />
     <Compile Include="TraceFileReader.cs" />
     <Compile Include="Utilities.cs" />
   </ItemGroup>
