+++ src/ImplTests/ImplTests.csproj
@@ -49,6 +49,9 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="LogFindTests.cs" />
+    <Compile Include="LogLineTests.cs" />
+    <Compile Include="LogTimeTests.cs" />
     <Compile Include="TraceFileReaderTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UtilitiesTests.cs" />
